+++ .changeset/famous-eggs-bathe.md
@@ -1,0 +1,5 @@
+---
+"@nestjs-modular/inject": patch
+---
+
+Make the code split into separate files for each use
